[Title] Edit the CI/CD System
[Desc]
The CMake's external project is too complex for the many module in the projects.
So, I decided using the system that is the "git submodules".

Then, the log project's needs the cmake's util functions that included the cmake's repo.
So, I practiced for the other project, then
add the cmake's repo to the log project.
[Type/Branch]
[Auther/Data]

## Type - fix / test / feature / refac / bug / study
## Branch -
## Auther
## Date
## Desc